Poll network status from SystemRegistry
This fixes an issue where you'd only get back a subset of the interface
state when calling Nerves.Network.status/1 and most importantly it
now tells you the IP address.

This also removes the direct status call to Nerves.Network.Interface,
since it would be good to remove that eventually.

This fixes #18.