Clarify that `Nerves.Network.teardown/1` does not do defaults
Any interfaces configured as a default will not be torn down by the
teardown function. This isn't readily clear to the user so a little
documentation is the next step. We can change this behaviour, but I
thought current documentation was the best first step.

Amos King @adkron <amos@binarynoggin.com>