Handle renew event inside dhcp manager
Originally the renw event was un-handled. When I started looking I
thought this was okay because it seemed like an informational event.
After looking further there is a possibility that the ip or other info
could change on a renew. Since that is the case we need to update our
configuration based on the information coming in. I thought of checking
and only notifying if the state actually changed, but that
would take more time than it would save since this shouldn't be a
frequent message.

Amos King @adkron <amos@binarynoggin.com>